backend/requirements.txt
# FastAPI and server
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5

# Face recognition
face-recognition>=1.3.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0

# Anti-spoofing
scipy>=1.7.0
scikit-image>=0.18.0

# Database
sqlalchemy>=1.4.0
aiosqlite>=0.17.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
